мапперы теперь модержатся не в объекте-синглтоне
